requirements.txt
Django==1.10
djangorestframework==3.5.3
django_cors_headers==1.3.1